Refactor chat mode initialization into dedicated AIChatModule
Extract lengthy initializeChatMode() function from saypi.index.js into a
dedicated AIChatModule class for better maintainability and encapsulation.
This follows the same pattern as UniversalDictationModule.

Changes:
- Create AIChatModule.ts with singleton pattern and encapsulated methods
- Move addVisualisations() and addLockButtons() into the module
- Simplify saypi.index.js by removing ~90 lines of chat-specific code
- Clean up unused imports in main index file
- Fix TypeScript warnings for unused parameters
